Change double to float.
Add float compare logic.
Also added src folder at Eclipse insistence.

+++ tests/TestService.java
@@ -2,13 +2,16 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
+import src.InputException;
+import src.Service;
+
 public class TestService {
 
     // test constructor that makes new objects
     @Test
     public void testServiceNew() {
         try {
-            new Service("Massage", 60.85);
+            new Service("Massage", 60.85f);
         } catch (Exception e) {
             fail(e.getMessage());
         }
@@ -19,7 +22,7 @@ public class TestService {
     @Test
     public void testService() {
         try {
-            new Service(1, "Massage", 60.85, 0);
+            new Service(1, "Massage", 60.85f, 0);
         } catch (Exception e) {
             fail(e.getMessage());
         }
@@ -28,7 +31,7 @@ public class TestService {
     @Test
     public void testServiceActive() {
       try {
-          Service service = new Service(1, "Massage", 60.85, 1);
+          Service service = new Service(1, "Massage", 60.85f, 1);
           boolean result = service.getStatus();
           assertEquals(true, result);
       } catch (InputException e) {
@@ -39,7 +42,7 @@ public class TestService {
     @Test
     public void testServiceInactive() {
       try {
-          Service service = new Service(1, "Massage", 60.85, 0);
+          Service service = new Service(1, "Massage", 60.85f, 0);
           boolean result = service.getStatus();
           assertEquals(false, result);
       } catch (InputException e) {
@@ -51,7 +54,7 @@ public class TestService {
     public void testServiceCodeNegative() {
         boolean thrown = false;
         try {
-            Service service = new Service(-1, "Massage", 60.85, 0);
+            Service service = new Service(-1, "Massage", 60.85f, 0);
         } catch (InputException e) {
             thrown = true;
         }
@@ -62,7 +65,7 @@ public class TestService {
     public void testServiceCodeTooLarge() {
         boolean thrown = false;
         try {
-            Service service = new Service(1000000, "Massage", 60.85, 0);
+            Service service = new Service(1000000, "Massage", 60.85f, 0);
         } catch (InputException e) {
             thrown = true;
         }
@@ -73,7 +76,7 @@ public class TestService {
     public void testServiceNameTooLong() {
         boolean thrown = false;
         try {
-            Service service = new Service(1, "TobleroneTobleroneToblerone", 60.85, 0);
+            Service service = new Service(1, "TobleroneTobleroneToblerone", 60.85f, 0);
         } catch (InputException e) {
             thrown = true;
         }
@@ -84,7 +87,7 @@ public class TestService {
     public void testServiceFeeNegative() {
         boolean thrown = false;
         try {
-            Service service = new Service(1, "Massage", -1.0, 0);
+            Service service = new Service(1, "Massage", -1.0f, 0);
         } catch (InputException e) {
             thrown = true;
         }
@@ -95,7 +98,7 @@ public class TestService {
     public void testServiceFeeTooLarge() {
         boolean thrown = false;
         try {
-            Service service = new Service(1, "Massage", 10000.0, 0);
+            Service service = new Service(1, "Massage", 10000.0f, 0);
         } catch (InputException e) {
             thrown = true;
         }
@@ -107,7 +110,7 @@ public class TestService {
     @Test
     public void testName() {
         try {
-            Service service = new Service("Massage", 0.0);
+            Service service = new Service("Massage", 0.0f);
             String result = service.getName();
             assertEquals("Massage", result);
         } catch (InputException e) {
@@ -118,7 +121,7 @@ public class TestService {
     @Test
     public void testCode() {
         try {
-            Service service = new Service(0, "", 0.0, 1);
+            Service service = new Service(0, "", 0.0f, 1);
             int result = service.getCode();
             assertEquals(0, result);
         } catch (InputException e) {
@@ -129,9 +132,9 @@ public class TestService {
     @Test
     public void testFee() {
         try {
-            Service service = new Service("", 60.85);
-            Double expectedResult = 60.85;
-            Double actualResult = service.getFee();
+            Service service = new Service("", 60.85f);
+            Float expectedResult = 60.85f;
+            Float actualResult = service.getFee();
             assertEquals(expectedResult, actualResult);
         } catch (InputException e) {
             fail(e.getMessage());
@@ -141,7 +144,7 @@ public class TestService {
     @Test
     public void testStatus() {
         try {
-            Service service = new Service(0, "", 60.85, 1);
+            Service service = new Service(0, "", 60.85f, 1);
             Boolean result = service.getStatus();
             assertEquals(true, result);
         } catch (InputException e) {
@@ -154,7 +157,7 @@ public class TestService {
     @Test
     public void testToString() {
         try {
-            Service service = new Service("Massage", 60.85);
+            Service service = new Service("Massage", 60.85f);
             String expectedResult = "0\tMassage\t$60.85";
             String actualResult = service.toString();
             assertEquals(expectedResult, actualResult);
@@ -169,8 +172,8 @@ public class TestService {
     public void testEquals() {
         boolean equal = false;
         try {
-            Service counseling = new Service(1, "Nutritionist", 60.85, 1);
-            Service sameCounseling = new Service(1, "Nutritionist", 60.85, 1);
+            Service counseling = new Service(1, "Nutritionist", 60.85f, 1);
+            Service sameCounseling = new Service(1, "Nutritionist", 60.85f, 1);
             equal = counseling.equals(sameCounseling);
         } catch (InputException e) {
             fail(e.getMessage());
@@ -182,8 +185,8 @@ public class TestService {
     public void testEqualsFails() {
         boolean equal = false;
         try {
-            Service counseling = new Service(1, "Massage", 60.85, 1);
-            Service massage = new Service(1, "Nutritionist", 60.85, 1);
+            Service counseling = new Service(1, "Massage", 60.85f, 1);
+            Service massage = new Service(1, "Nutritionist", 60.85f, 1);
             equal = counseling.equals(massage);
         } catch (InputException e) {
             fail(e.getMessage());
